[FAB-1335] Verify the setting of environ vars
This allows for the testing of the setting of environment
variables. It also adds some tests that can be used as
smoke tests for executing the functionality of the behave
tests to make sure things are working as expected.

This can be used by CI when code is pushed to this repository
for verifying changes without executing the full suite of tests.

+++ feature/steps/compose_util.py
@@ -218,5 +218,7 @@ class Composition:
                 subprocess.call(['docker', 'rm', '-f', container], env=env)
 
         # Need to remove the chaincode images: docker rmi -f $(docker images | grep "example.com-" | awk '{print $3}')
-        cmd = ['docker images | grep ".example.com-" | awk \'{print $3}\' | xargs docker rmi']
-        subprocess.call(cmd, shell=True, env=env)
+        retVal = subprocess.call(['docker images | grep ".example.com-"'], env=env, shell=True)
+        if retVal != 1:
+            cmd = ['docker images | grep ".example.com-" | awk \'{print $3}\' | xargs docker rmi']
+            subprocess.call(cmd, shell=True, env=env)
